Add --own-name=org.kde.*, set TMPDIR to fix status notifier icon
The status notifier protocol seems to require ownership of the D-Bus
name org.kde.StatusNotifierItem-PID-ID, and org.kde.* is the strictest
wildcard supported by Flatpak that includes names of that format.  The
icon is written to a PNG file in TMPDIR, so we need to set that to a
path accessible from outside the sandbox.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>